Extract CrossFade into Transitions
