+++ src/views/Transition/Transitions/index.js
@@ -1,5 +1,7 @@
 import SharedElement from './SharedElement';
+import CrossFade from './CrossFade';
 
 export default {
   SharedElement,
+  CrossFade,
 }
